ParameterVector::push_back()
Also move constructor to parameter_vector.C so we can remove the
parameter_pointer.h dependency from parameter_vector.h
