+++ src/systems/parameter_vector.C
@@ -20,6 +20,7 @@
 // C++ Includes   -----------------------------------
 
 // Local Includes -----------------------------------
+#include "libmesh/parameter_pointer.h"
 #include "libmesh/parameter_vector.h"
 
 namespace libMesh
@@ -27,6 +28,17 @@ namespace libMesh
 
 // ------------------------------------------------------------
 // ParameterVector implementation
+
+
+
+ParameterVector::ParameterVector(const std::vector<Number *> &params)
+  : _is_shallow_copy(false)
+{
+  _params.reserve(params.size());
+
+  for (unsigned int i=0; i != params.size(); ++i)
+    _params.push_back(new ParameterPointer<Number>(params[i]));
+}
